fix: cumulative utilisation fixed, tested, and working

+++ contracts/interfaces/ILiquidationModule.sol
@@ -53,7 +53,7 @@ interface ILiquidationModule is IBasePerpMarket {
      */
     function getRemainingLiquidatableSizeCapacity(
         uint128 marketId
-    ) external view returns (uint128 maxLiquidatableCapacity, uint128 remainingCapacity, uint64 lastLiquidationTime);
+    ) external view returns (uint128 maxLiquidatableCapacity, uint128 remainingCapacity, uint128 lastLiquidationTime);
 
     /**
      * @notice Returns whether a position owned by `accountId` can be flagged for liquidated.
